major fix, entity con chiave primaria Long, major fix relazioni, minor fix nomi delle classi e references

+++ src/main/java/com/gol/ants_quests/hibernate/entities/EsitoQuest.java
@@ -24,7 +24,7 @@ public class EsitoQuest extends GenericEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idEstQst;
+    private Long idEstQst;
     private Date dataEsecuzione;
     private String punteggio;
     private String tempo;
